Move javascript function to a js file
